Live migration for the read marker in case it failed on the update
The current migration only works if the user read in the same exact second
as the comment was posted (author only). So we set the read marker to -1
for all users and in case of -1 we calculate the marker on the next request,
because it is not possible in a programmatic way with a reasonable O(1) or O(n)
but requires one query per user&chat.

Signed-off-by: Joas Schilling <coding@schilljs.com>