Fix pending forced reconnection after WebSocket reconnection
When the WebSocket is disconnected from the external signaling server it
is automatically reconnected to it; if a forced reconnection was queued
while the WebSocket was disconnected it was supposed to be performed as
soon as it was reconnected. However, "_forceReconnect" tracks whether
there is a pending forced reconnection, and as it was set to false when
"Signaling.Standalone.connect()" was called the queued forced
reconnection never happened after a WebSocket reconnection (which caused
the previous session to be resumed instead of a new one being created).

"_forceReconnect" is only used to force the reconnection with a new
session after receiving the hello response from the server, and it is
set to false when the forced reconnection is properly done, so it can be
safely removed from "connect()".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>